Renames
Teeth_Spike_Doors -> Slamming_Doors
