+++ TombEngine/Objects/TR1/Trap/SlammingDoors.h
