+++ TombEngine/Scripting/Internal/TEN/Objects/ObjectIDs.h
@@ -364,7 +364,7 @@ The following constants are inside ObjID.
 	EXPLOSION
 	DAMOCLES_SWORD
 	ELECTRIC_CLEANER
-	TEETH_SPIKE_DOOR
+	SLAMMING_DOORS
 	PUZZLE_ITEM1
 	PUZZLE_ITEM2
 	PUZZLE_ITEM3
@@ -1532,7 +1532,7 @@ static const std::unordered_map<std::string, GAME_OBJECT_ID> kObjIDs {
 	{ "EXPLOSION", ID_EXPLOSION },
 	{ "DAMOCLES_SWORD", ID_DAMOCLES_SWORD },
 	{ "ELECTRIC_CLEANER", ID_ELECTRIC_CLEANER },
-	{ "TEETH_SPIKE_DOOR", ID_TEETH_SPIKE_DOOR},
+	{ "SLAMMING_DOORS", ID_SLAMMING_DOORS },
 	{ "PUZZLE_ITEM1", ID_PUZZLE_ITEM1 },
 	{ "PUZZLE_ITEM2", ID_PUZZLE_ITEM2 },
 	{ "PUZZLE_ITEM3", ID_PUZZLE_ITEM3 },
